started implementation of watershed seg

+++ src/main/java/pt/unl/itqb/imagej/parameters/SegmentationParameters.java
@@ -2,13 +2,13 @@ package pt.unl.itqb.imagej.parameters;
 
 public class SegmentationParameters {
     private int peakmindist;
-    private int peakminheight;
+    private double peakminheight;
     private int peakminedge;
     private int maxpeaks;
 
     public SegmentationParameters() {
         peakmindist = 5;
-        peakminheight =  5;
+        peakminheight =  5.0;
         peakminedge = 10;
         maxpeaks = 10000;
     }
@@ -21,11 +21,11 @@ public class SegmentationParameters {
         return this.peakmindist;
     }
 
-    public void setPeakminheight(int minheight) {
+    public void setPeakminheight(double minheight) {
         this.peakminheight = minheight;
     }
 
-    public int getPeakminheight() {
+    public double getPeakminheight() {
         return this.peakminheight;
     }
